requirements.txt
numpy
websocket-client
matplotlib
